Fix three offload bugs found bringing EP up on real hardware
* expert_cache: the pinned host bank was allocated with
  torch.empty_like(host, pin_memory=True), no device. Diversion runs
  inside the model's construction device context, whose torch_function
  mode fills in device=cuda for factory calls that do not name one, so
  this asked for pinned memory *on the GPU* and failed as "CUDA error:
  out of memory" regardless of free host RAM. Probing the same box
  outside a device context pinned 36 GiB happily, which is what made
  the earlier "WSL2 pinned-memory ceiling" reading wrong. Names
  device="cpu" explicitly.

* PLE disk (staged sync): masked_fill_ has no kernel for
  float8_e4m3fn, the table dtype, so the TP mask added for expert
  parallelism crashed every forward. Applies the fill through a
  same-width uint8 view instead (an all-zero byte pattern is +0.0 in
  e4m3/e5m2), keeping the op fixed-shape and CUDA-graph capturable
  rather than falling back to boolean-mask indexing. The TP all-reduce
  that follows now runs in bfloat16 for the same dtype-support reason;
  it is lossless here because every e4m3 value is representable in
  bfloat16 and exactly one rank contributes each row.

* fused copy: bank rows had to be a multiple of 16 bytes, but NVFP4
  banks per-expert fp32 scalar scales (4-byte rows), and one such bank
  disqualified the whole plan -- dropping every bank onto the legacy
  host-count copy, which host-synchronizes per bank per MoE layer per
  step. On real hardware that alone made decode slow enough for vLLM's
  own sample_tokens RPC to time out. The copy is PCIe-bound, so the
  unit is now 4 bytes (int32) and the alignment requirement follows.
  A test pins the invariant so a future bank cannot silently
  reintroduce the legacy path.

+++ vllm/models/qwen4_exp/nvidia/ple_layer.py
@@ -608,7 +608,14 @@ class Qwen4ExpNGramEmbedding(nn.Module):
         # free). Skipped at tp_size==1, where the mask is always all-False
         # anyway but the collective would still cost a no-op sync.
         if self.ngram_embedding.tp_size > 1:
-            rows = tensor_model_parallel_all_reduce(rows)
+            # Reduce in a wider dtype: the rows are the table's fp8, which
+            # has no all-reduce path we can count on. The round trip is
+            # lossless here -- every e4m3 value is exactly representable in
+            # bfloat16, and each row is contributed by exactly one rank with
+            # zeros from the others, so the sum is that one value unchanged.
+            rows = tensor_model_parallel_all_reduce(rows.to(torch.bfloat16)).to(
+                rows.dtype
+            )
         return rows.flatten(-2)
 
     def gather_disk_rows(self, ngram_ids: torch.Tensor, output: torch.Tensor) -> None:
